% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_functions.R
\name{get_trips}
\alias{get_trips}
\title{Pull trips from shl, fhv or yellow records for a given date range}
\usage{
get_trips(service, dt_start = NULL, dt_end = NULL, features = NULL,
  dt_convert = T)
}
\arguments{
\item{service}{the service you want in quotes: 'fhv'.}

\item{dt_start}{start date of range.}

\item{dt_end}{end date of range.}

\item{features}{a character vector of desired variables.}

\item{dt_convert}{do you want dates to be automatically converted, defaults to T.}
}
\description{
This function allows you to access rbin files on the shared ride to query quickly trip data from any of our services.
}
\examples{
get_trips("fhv",dt_start = "2017-06-01", dt_end = "2017-06-01", features = c("hack","plate","pudt","dodt"))
}
\keyword{get}
