% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trip_functions.R
\name{get_trips_shared}
\alias{get_trips_shared}
\title{Access shared trips from fhv records for a given date range}
\usage{
get_trips_shared(dt_start = NULL, dt_end = NULL, features = NULL,
  merge_ent = F, odbc_con = NULL, query = NULL, left_key = NULL,
  right_key = NULL, join_type = NULL)
}
\arguments{
\item{dt_start}{start date of range.}

\item{dt_end}{end date of range.}

\item{features}{a character vector of desired variables.}

\item{merge_ent}{if True merges against entity automatically, defaults to FALSE.}

\item{odbc_con}{for accessing sql tables in function, pass your odbc connection string.}

\item{query}{a character string encompassing the sql query.}

\item{left_key}{the left join key in the trip records.}

\item{right_key}{the right join key in entity.}

\item{join_type}{the type of join desired, currently experimental, defaults to left.}

\item{hack_fil}{a character vector of desired hacks to filter on.}
}
\description{
This function allows you to access rbin files on the shared drive to query quickly share trip data. It also provides access to quick joins on the share data and sql backend.
}
\examples{
get_trips_shared(dt_start = "2017-06-01", dt_end = "2017-06-03", features = c("hack","plate","pudt","dodt", "baseDisp"),merge_ent = T,odbc_con = "datawarehouse",query = "select ltrim(rtrim(lic_no)) as lic_no,trade_nam as company from tlc_camis_entity where lic_code like 'BAS'",left_key = "baseDisp",right_key = "lic_no")
}
\keyword{get}
